Add the start of the ncmec class and related helper classes
Currently includes the class for a ncmec IP object (checkuser data) and the start of the larger wrapper class ncmec
The current class doesn't do a full report but is able to check server status and retract a report.

Currently full reports should still use the main processor, those will be next to add after retraction and status code is updated to use this.